feat: complete native binding and standardize public API contracts
Add native DATE and DECIMAL binders with checked binary encoding and
independent batch snapshots. Keep text input under explicit _string names.

Correct scalar NULL and overflow fallbacks, use size_t column indices,
and add status-based statement fetching. Give string and byte getters
separate capacity, required-size and NULL outputs without silent truncation.

Make Smart-LOB handles opaque with explicit client ownership and semantic
accessors. Migrate consumers and document the experimental API changes.

Cover boundary values, allocation failures, buffer atomicity and ownership
with regression tests, and exercise public native binders in wire fixtures.

+++ src/sqli_protocol_internal.h
@@ -7,7 +7,7 @@ bool sqli_protocol_has_buffered_data(sqli_conn_t *conn, int fd);
 sqli_status sqli_send_stmt_close_cursor(sqli_conn_t *conn, int stmt_id);
 void sqli_stmt_close_release(sqli_conn_t *conn, int stmt_id);
 int64_t sqli_decode_ifx_int8(const uint8_t *buf, bool *is_null);
-sqli_status sqli_fetchblob_materialize(sqli_result_t *result, int col_index,
+sqli_status sqli_fetchblob_materialize(sqli_result_t *result, size_t col_index,
                                        uint8_t **blob_buf, size_t *blob_len);
 bool sqli_is_lob_like_type(uint8_t type);
 bool sqli_is_legacy_lob_type(uint8_t type);
